plugin is code complete
