+++ solr/core/src/java/org/apache/solr/cluster/placement/impl/ReplicaPlacementImpl.java
@@ -67,7 +67,7 @@ class ReplicaPlacementImpl implements ReplicaPlacement {
     List<ReplicaPosition> replicaPositions = new ArrayList<>(replicaPlacementSet.size());
     int index = 0; // This really an arbitrary value when adding replicas and a possible source of core name collisions
     for (ReplicaPlacement placement : replicaPlacementSet) {
-      replicaPositions.add(new ReplicaPosition(placement.getShardName(), index++, ReplicaImpl.toCloudReplicaType(placement.getReplicaType()), placement.getNode().getName()));
+      replicaPositions.add(new ReplicaPosition(placement.getShardName(), index++, SimpleClusterAbstractionsImpl.ReplicaImpl.toCloudReplicaType(placement.getReplicaType()), placement.getNode().getName()));
     }
 
     return replicaPositions;
